+++ src/asldro/filters/ground_truth_loader.py
@@ -36,6 +36,14 @@ class GroundTruthLoaderFilter(BaseFilter):
     :type 'parameters': dict
     :param 'segmentation': dictionary containing key-value pairs corresponding
         to tissue type and label value in the "seg_label" volume.
+    :param 'image_override': (optional) dictionary containing single-value override values
+    for any of the 'image' that are loaded. The keys must match the quantity name
+    defined in 'quantities'.
+    :param 'image_override': dict
+    :param 'parameter_override': (optional) dictionary containing single-value override values
+    for any of the 'parameters' that are loaded. The keys must match the key defined in
+    'parameters'.
+    :param 'parameter_override': dict
 
     **Outputs**
 
@@ -46,10 +54,13 @@ class GroundTruthLoaderFilter(BaseFilter):
     corresponding to a 3D/4D subset of the nifti input (split along the 5th
     dimension). The data types of images will be the
     same as those input EXCEPT for a quantity labelled "seg_label"
-    which will be converted to a uint16 data type.  The keys-value pars in
-    the input 'parameters' will also be destructured and piped through
-    to the output' For example:
-
+    which will be converted to a uint16 data type.
+    If 'override_image' is defined, the corresponding 'image' will be set to the overriding
+    value before being output.
+    If 'override_parameters' is define, the corresponding parameter will be set to the
+    overriding value before being output.
+    The keys-value pairs in the input 'parameters' will also
+    be destructured and piped through to the output, for example:
     :param 't1': volume of T1 relaxation times
     :type 't1': NiftiImageContainer
     :param 'seg_label': segmentation label mask corresponding to different tissue types.
@@ -82,6 +93,8 @@ class GroundTruthLoaderFilter(BaseFilter):
     KEY_QUANTITIES = "quantities"
     KEY_QUANTITY = "quantity"
     KEY_MAG_STRENGTH = "magnetic_field_strength"
+    KEY_IMAGE_OVERRIDE = "image_override"
+    KEY_PARAMETER_OVERRIDE = "parameter_override"
 
     def __init__(self):
         super().__init__("GroundTruthLoader")
@@ -107,6 +120,14 @@ class GroundTruthLoaderFilter(BaseFilter):
                 # squeeze the 4th dimension if there is only one time-step
                 image_data = np.squeeze(image_data, axis=3)
 
+            # If we have a corresponding 'image_override' value, update the
+            # 'image_data' with that value
+            if (
+                self.KEY_IMAGE_OVERRIDE in self.inputs
+                and quantity in self.inputs[self.KEY_IMAGE_OVERRIDE]
+            ):
+                image_data.fill(self.inputs[self.KEY_IMAGE_OVERRIDE][quantity])
+
             # If we have a segmentation label, round and squash the
             # data to uint16 and update the NIFTI header
             metadata = {}
@@ -129,8 +150,15 @@ class GroundTruthLoaderFilter(BaseFilter):
                 metadata=metadata,
             )
             self.outputs[quantity] = new_image_container
+
+        # Get the parameter_override dictionary (empty dict if it doesn't exist)
+        overrides = (
+            self.inputs[self.KEY_PARAMETER_OVERRIDE]
+            if self.KEY_PARAMETER_OVERRIDE in self.inputs
+            else {}
+        )
         # Pipe through all parameters
-        self.outputs = {**self.outputs, **self.inputs[self.KEY_PARAMETERS]}
+        self.outputs = {**self.outputs, **self.inputs[self.KEY_PARAMETERS], **overrides}
 
     def _validate_inputs(self):
         """Checks that the inputs meet their validation criteria.
@@ -139,8 +167,14 @@ class GroundTruthLoaderFilter(BaseFilter):
         'units': list[str]
         'segmentation': dict
         'parameters': dict
-        The number of 'units' and 'quantities' should be equal
-        The size of the 5th dimension of the image must equal the number of 'quantities'
+        'image_override': dict (optional)
+        'parameter_override': dict (optional)
+        The number of 'units' and 'quantities' should be equal.
+        The size of the 5th dimension of the image must equal the number of 'quantities'.
+        If 'image_override' is present, must be a dict, each of the keys must be a string
+        and the values a float or int type. The key must match an entry in 'quantities'.
+        If 'parameter_override' is present, must be a dict, each of the keys must be a
+        string and the values a float or int type.
         """
         input_validator = ParameterValidator(
             parameters={
@@ -174,6 +208,34 @@ class GroundTruthLoaderFilter(BaseFilter):
 
         if len(self.inputs["units"]) != len(self.inputs["quantities"]):
             raise FilterInputValidationError(
-                f"{self} filter requires an input 'units' which is the same length as the input"
-                "quantities"
+                f"{self} filter requires an input 'units' which is the same length as the input "
+                "'quantities'"
             )
+
+        for input_key in [self.KEY_IMAGE_OVERRIDE, self.KEY_PARAMETER_OVERRIDE]:
+            if input_key in self.inputs:
+                # Check the value is a dictionary
+                if not isinstance(self.inputs[input_key], dict):
+                    raise FilterInputValidationError(
+                        f"{input_key} must be a dictionary, is {self.inputs[input_key]}"
+                    )
+                # Check all of the keys are strings and the values are int or float
+                for key, value in self.inputs[input_key].items():
+                    if not isinstance(key, str):
+                        raise FilterInputValidationError(
+                            f"All keys in the {input_key} dictionary must be strings. "
+                            f"{key} is not."
+                        )
+                    if not isinstance(value, (int, float)):
+                        raise FilterInputValidationError(
+                            f"Values in the {input_key} dictionary must be int or float. "
+                            f"The value for {key} ({value}) is not."
+                        )
+
+        if self.KEY_IMAGE_OVERRIDE in self.inputs:
+            # Check all of the keys are present in self.inputs['quantities']
+            for key, value in self.inputs[self.KEY_IMAGE_OVERRIDE].items():
+                if key not in self.inputs[self.KEY_QUANTITIES]:
+                    raise FilterInputValidationError(
+                        f"{key} is not in the input 'quantities' list"
+                    )